requirements.txt
opencv-python
numpy
requests
torch
supervision
openai>=1.3.0
python-dotenv
